Move dllogger to separate file and check import
